Fix image name option for encrypt-gce-image
When no --encrypted_image_name option is passed
the encrypt command now generates a name for the
encrypted image rather than failing silently. It
uses the same log that update-gce-image uses to
generate a same (but the session id will change
so this won't cause collisions).

Removes the validate option from encrypt-gce-image
because it wasn't doing anything.

Moves gce specific functions into gce_service.py

Testing: encrypted an image without the
encrypted_image_name option and then ran
update-gce-image on that encrypted image
