+++ brkt_cli/update_encrypted_gce_image_args.py
@@ -15,13 +15,6 @@ def setup_update_gce_image_args(parser):
         required=False
     )
     parser.add_argument(
-        '--no-validate',
-        dest='validate',
-        action='store_false',
-        default=True,
-        help="Don't validate images"
-    )
-    parser.add_argument(
         '--zone',
         help='GCE zone to operate in',
         dest='zone',
